+++ asm/melee/ft/chara/ftnana.s
