+++ asm/melee/ft/chara/ftpopo.s
