+++ asm/melee/lb/lb_00CE.s
