+++ asm/melee/lb/lb_00F9.s
