+++ asm/melee/lb/lb_0192.s
