+++ asm/sysdolphin/baselib/hsd_40FF.s
