+++ src/melee/ft/chara/ftCLink/ftCLink_Init.c
