+++ src/melee/ft/chara/ftCaptain/ftCaptain_Init.c
