+++ src/melee/ft/chara/ftMario/ftMario_Init.c
