+++ src/melee/ft/chara/ftPopo/fticeclimber1.h
