+++ src/melee/ft/chara/ftPopo/ftpopo_init.c
