+++ src/melee/ft/chara/ftZelda/ftZelda_Init.c
