+++ src/melee/ft/ftdata.c
@@ -23,9 +23,6 @@
 #include "ftGameWatch/ftgamewatch.h"
 #include "ftGanon/ftganon.h"
 #include "ftGigaKoopa/ftgigakoopa.h"
-#include "ftIceClimber/fticeclimber.h"
-#include "ftIceClimber/fticeclimber1.h"
-#include "ftIceClimber/fticeclimber2_nana.h"
 #include "ftKirby/ftkirby.h"
 #include "ftKoopa/ftkoopa.h"
 #include "ftLink/ftlink.h"
@@ -36,11 +33,14 @@
 #include "ftMasterHand/ftMasterHand_01.h"
 #include "ftMasterHand/ftMasterHand_02.h"
 #include "ftMewtwo/ftmewtwo.h"
+#include "ftNana/fticeclimber2_nana.h"
 #include "ftNess/ftNess.h"
 #include "ftPeach/ftpeach.h"
 #include "ftPichu/ftpichu.h"
 #include "ftPikachu/ftpikachu.h"
 #include "ftPikachu/ftpikachu1.h"
+#include "ftPopo/fticeclimber.h"
+#include "ftPopo/fticeclimber1.h"
 #include "ftPurin/ftpurin.h"
 #include "ftSamus/ftsamus1.h"
 #include "ftSamus/ftsamus3.h"
